LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-in-report-only-ignored-expected.txt

CONSOLE MESSAGE: The Content Security Policy directive 'frame-ancestors' is ignored when delivered in a report-only policy.
Tests that loading a page in an <iframe> with a report-only Content Security Policy "frame-ancestors 'none'" is allowed. This test PASSED if you see the word PASS below. Otherwise, it FAILED.



--------
Frame: '<!--frame1-->'
--------
PASS
